Fixes issues with Remoting CLI and Security settings after initial setup.
Now only the plugins installation should be done in groovy because currently they are only available after a
reboot (or ether not then!).

Also is the jenkins currently open for everyone to see but only for admin to administrate.

Furter the host name should be configured for the initial setup (get name from Docker?).